refactor: address review, rm duplicated primitives
