+++ crates/rbuilder/src/live_builder/config.rs
@@ -36,7 +36,7 @@ use crate::{
     },
     mev_boost::{
         bloxroute_grpc, BLSBlockSigner, MevBoostRelayBidSubmitter, MevBoostRelaySlotInfoProvider,
-        RelayClient,
+        RelayClient, RelayConfig, RelaySubmitConfig,
     },
     provider::StateProviderFactory,
     roothash::RootHashContext,
@@ -54,7 +54,7 @@ use ethereum_consensus::{
 };
 use eyre::Context;
 use lazy_static::lazy_static;
-use rbuilder_primitives::mev_boost::{MevBoostRelayID, RelayConfig, RelayMode, RelaySubmitConfig};
+use rbuilder_primitives::mev_boost::{MevBoostRelayID, RelayMode};
 use reth_chainspec::{Chain, ChainSpec, NamedChain};
 use reth_db::DatabaseEnv;
 use reth_node_api::NodeTypesWithDBAdapter;
